Main functionality and installation support
