+++ lint_all/pyproject.toml
@@ -1,0 +1,34 @@
+[project]
+name = "lint_all"
+version = "0.1.0"
+description = "A meta linter to run all linters"
+dependencies = [
+#        "cpplint",
+        "gitdb",
+        "GitPython",
+        "mypy",
+        "pyenchant",
+#        "pylint",
+        "types-mock",
+#       "types-protobuf",
+#        "types-redis",
+#        "types-requests",
+#        "types-setuptools",
+#        "types-six",
+#        "types-pyyaml",
+        "pyyaml",]
+requires-python = ">=3.8"
+readme = "README.md"
+
+[[project.authors]]
+name = "Shir Granot Peled"
+email = "shirpeled+lintall@gmail.com"
+
+[project.scripts]
+lint_all = "lint_all:parse_args_and_run"
+
+[project.license]
+text = "Apache-2.0"
+
+[tool.setuptools.package-data]
+"*" = ["*.*"] # <-- this changed
